# docs/swagger.yaml
definitions:
...
  model.ParamLogin:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  model.ParamSignUp:
    properties:
      password:
        type: string
      repassword:
        type: string
      unit:
        type: string
      username:
        type: string
      verifycode:
        type: string
    required:
    - password
    - repassword
    - unit
    - username
    - verifycode
    type: object
...
  respcode.ResponseData:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: object
    type: object
info:
  contact: {}
  description: 交通一张图
  title: 交通一张图后端系统
  version: 1.0.0
paths:
  /api/v1/cache/hit/{infotype}/{year}/{level}:
    get:
      consumes:
      - application/json
      description: 检查缓存中是否有请求的值 有就返回没有将请求转发
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 查询类型
        in: path
        name: infotype
        required: true
        type: string
      - description: '查询年份 格式: 202X '
        in: path
        name: year
        required: true
        type: string
      - description: 查询等级
        in: path
        name: level
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 检查缓存命中接口
      tags:
      - 缓存相关接口
...
  /api/v1/data/menus:
    get:
      consumes:
      - application/json
      description: 用于前端渲染数据管理侧边栏
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取数据导航栏
      tags:
      - 用户相关api
  /api/v1/home/menus:
    get:
      consumes:
      - application/json
      description: 用于前端渲染用户家目录侧边栏
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取导航栏数据
      tags:
      - 用户相关api
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 首页登录
      parameters:
      - description: 用户登录信息
        in: body
        name: 登录信息
        required: true
        schema:
          $ref: '#/definitions/model.ParamLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 处理登录请求
      tags:
      - 登陆api
  /api/v1/signup:
    get:
      consumes:
      - application/json
      description: 申请一个注册码 重复请求上一个就会失效 重启后端服务也会失效
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 申请一个注册码
      tags:
      - 注册api
    post:
      consumes:
      - application/json
      description: 首页注册
      parameters:
      - description: 用户注册信息
        in: body
        name: 注册信息
        required: true
        schema:
          $ref: '#/definitions/model.ParamSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respcode.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  type: string
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 处理注册请求
      tags:
      - 注册api
...
swagger: "2.0"